% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:BHM.R
\name{berrymethod.calibrate}
\alias{berrymethod.calibrate}
\title{Simulation for BHM Method}
\usage{
berrymethod.calibrate(
  cohortsize,
  ntype = 4,
  p.true,
  p.null,
  p.target,
  ntrial = 10000,
  n.burnin = 2000,
  n.chain = 1,
  n.iter = 30000,
  mu.par,
  FWER,
  precision = 1e-04,
  fwer.threshold = 0.001
)
}
\arguments{
\item{cohortsize}{vector format, sample size for each cohort per disease type}

\item{ntype}{number of disease types/subgroups}

\item{p.true}{the vector of true response rate in each disease type}

\item{p.null}{the vector of null hypothesis rate in each type}

\item{p.target}{the vector of alternative hypothesis rate in each type}

\item{ntrial}{the number of simulated trials}

\item{n.burnin}{control prameters of MCMC sampling using jags}

\item{n.chain}{control prameters of MCMC sampling using jags}

\item{n.iter}{control prameters of MCMC sampling using jags}

\item{mu.par}{the mean parameter of mu's prior distribution}

\item{FWER:}{the pre-specified family wise error rate for efficacy cutoff calculation#}

\item{precision:}{the grid space controlled in the grid search, can be refined       #}

\item{fwer.threshold:}{the acceptable threshold for the difference between           #}
}
\value{
probability of reject
}
\description{
Simulation for BHM Method
}
