% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Optimal_2stage_design.R
\name{cal.N.2stg}
\alias{cal.N.2stg}
\title{This function use binomial exact test to calculate the sample size.}
\usage{
cal.N.2stg(p, alpha, beta)
}
\arguments{
\item{p}{probability vector, p\link{1} is the null hypothesis probability, p\link{2} is the alternative hypothesis probability}

\item{alpha}{type I error}

\item{beta}{type II error}
}
\value{
sample size
}
\description{
This function use binomial exact test to calculate the sample size.
}
