% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:CBHM.R
\name{CBHMdesign.calibrate}
\alias{CBHMdesign.calibrate}
\title{Get efficacy cutoff}
\usage{
CBHMdesign.calibrate(
  cohortsize,
  ntype = 4,
  p.true,
  p.null,
  p.target,
  ntrial = 5000,
  n.burnin,
  n.chain,
  n.iter,
  mu.par,
  FWER,
  precision = 0.001,
  fwer.threshold = 0.001,
  a,
  b
)
}
\arguments{
\item{cohortsize}{vector format, sample size for each cohort per disease type}

\item{ntype}{number of disease types/subgroups}

\item{p.true}{the vector of true response rate in each disease type}

\item{p.null}{the vector of null hypothesis rate in each type}

\item{p.target}{the vector of alternative hypothesis rate in each type}

\item{ntrial}{the number of simulated trials}

\item{n.burnin}{control prameters of MCMC sampling using jags}

\item{n.chain}{control prameters of MCMC sampling using jags}

\item{n.iter}{control prameters of MCMC sampling using jags}

\item{mu.par}{the mean parameter of mu's prior distribution}

\item{FWER}{the pre-specified family wise error rate for efficacy cutoff calculation#}

\item{a}{the tuning parameter that characterize the relationship between test       #}

\item{b}{the tuning parameter that characterize the relationship between test       #}

\item{precision:}{the grid space controlled in the grid search, can be refined       #}

\item{fwer.threshold:}{the acceptable threshold for the difference between           #}
}
\value{
efficacy cutoff
}
\description{
Get efficacy cutoff
}
