% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:CBHM.R
\name{CBHMdesign}
\alias{CBHMdesign}
\title{CBHMdesign() Function}
\usage{
CBHMdesign(
  cohortsize,
  ntype = 4,
  p.true,
  p.null,
  p.target,
  ntrial = 1,
  n.burnin,
  n.chain,
  n.iter,
  mu.par,
  eff.ref,
  a,
  b,
  r.all
)
}
\arguments{
\item{cohortsize}{vector format, sample size for each cohort per disease type}

\item{ntype}{number of disease types/subgroups}

\item{p.null}{the vector of null hypothesis rate in each type}

\item{p.target}{the vector of alternative hypothesis rate in each type}

\item{ntrial}{number of simulated trials needs to be 1 here, repeated simulations are done outside}

\item{n.burnin}{control prameters of MCMC sampling using jags}

\item{n.chain}{control prameters of MCMC sampling using jags}

\item{n.iter}{control prameters of MCMC sampling using jags}

\item{mu.par}{the mean parameter of mu's prior distribution}

\item{eff.ref}{the efficacy evaluation criteria, calibrated through simulations}

\item{a}{the tuning parameter that characterize the relationship between test       #}

\item{b}{the tuning parameter that characterize the relationship between test       #}

\item{r.all}{number of responses}
}
\value{
reponses and sample sizes
}
\description{
CBHMdesign() Function
}
