% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Optimal_1stage_design.R
\name{diff0.emp.1stg}
\alias{diff0.emp.1stg}
\title{This function finds alphastar}
\usage{
diff0.emp.1stg(K, g, alphastar, r, n, p, alpha, nsim)
}
\arguments{
\item{K}{number of indications}

\item{g}{the prior probability of H0: that the strata are completely correlated}

\item{alphastar}{adjusted alpha}

\item{r}{response probability for inactive drug}

\item{n}{number of patients}

\item{p}{probability vector, p\link{1} is the null hypothesis probability, p\link{2} is the alternative hypothesis probability}

\item{alpha}{global type I error}

\item{nsim}{number of simulation used for type I and type II calculation}
}
\value{
return alphastar
}
\description{
This function finds alphastar
}
