% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Optimal_2stage_design.R
\name{diff0.emp.2stg}
\alias{diff0.emp.2stg}
\title{This function finds alphastar}
\usage{
diff0.emp.2stg(alphastar, K, alpha0, beta0, n12pts, p0, p1, nsim, alpha)
}
\arguments{
\item{alphastar}{adjusted alpha}

\item{K}{number of indications}

\item{alpha0}{the type I error parameter for stage1}

\item{beta0}{the type II error parameter for stage1}

\item{n12pts}{total sample size for each indication}

\item{p0}{probability vector of the null hypothesis}

\item{p1}{probability vector of the alternative hypothesis}

\item{nsim}{number of simulation used for type I and type II calculation}

\item{alpha}{type I error}
}
\value{
return alphastar
}
\description{
This function finds alphastar
}
