% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Simon_two_stage.R
\name{EN}
\alias{EN}
\title{Expected N}
\usage{
EN(r1, n1, p, n)
}
\arguments{
\item{r1}{number of responses at IA}

\item{n1}{number of patients at IA}

\item{p}{probability vector}

\item{n}{number of patients}
}
\value{
Expected sample size
}
\description{
Expected N
}
