% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:EXNEX.R
\name{exnex}
\alias{exnex}
\title{Main function for EXNEX Method}
\usage{
exnex(
  num.sim = 1,
  Q.exnex = rep(0.8, 6),
  p0 = rep(0.2, 6),
  H0 = rep(0.2, 6),
  H1 = rep(0.4, 6),
  nk = matrix(rep(c(12, 24), 6), nrow = 2),
  K = 6,
  v1,
  Nexch = 2,
  Nmix = 3,
  pMix = c(0.5, 0, 0.5),
  dirichprior = c(1, 1, 1),
  dirichprior.check = TRUE,
  mu.mean = c(-1.735, 0.847),
  mu.prec = c(0.146, 0.266),
  tau.HN.scale = c(1, 1),
  nex.mean = rep(-1.734, 6),
  nex.prec = rep(0.128, 6),
  n.burnin = 1000,
  n.chain = 1,
  n.iter = 8000,
  r.all
)
}
\arguments{
\item{num.sim}{number of simulated trials needs to be 1 here, repeated simulations are done outside}

\item{Q.exnex}{efficacy cutoff for each cohort}

\item{p0}{the vector of true response rate in each disease type}

\item{H0}{the vector of null hypothesis rate in each type}

\item{H1}{the vector of alternative hypothesis rate in each type}

\item{nk}{sample size for each cohort at each analysis (can be different). row is for analysis#}

\item{v1}{efficacy cutoff}

\item{Nexch}{the no. of exchangeability distributions}

\item{Nmix}{==Nexch+1. no. of total candidate prior dists for theta_j.}

\item{pMix}{the probs of each strata following each dist.}

\item{dirichprior}{parameter for prior distribution specified in exnex.dirich.txt. but we do not have this file.}

\item{dirichprior.check}{if TRUE, will use exnex.dirich.txt to specify bayesian model}

\item{mu.mean}{mean of prior dist of mu in exchangeable part}

\item{mu.prec}{std of prior dist of mu in exchangeable part}

\item{tau.HN.scale}{parameter to derive prior dist of}

\item{nex.mean}{to derive prior dist of theta_j under non-exchangeability}

\item{nex.prec}{to derive prior dist of theta_j under non-exchangeability}

\item{n.burnin}{control prameters of MCMC sampling using jags}

\item{n.chain}{control prameters of MCMC sampling using jags}

\item{n.iter}{control prameters of MCMC sampling using jags}

\item{r.all}{number of responses}

\item{k}{no. of cohorts/indications}
}
\value{
probability of reject
}
\description{
Main function for EXNEX Method
}
