% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Simon_Bayes_design.R
\name{find_postk}
\alias{find_postk}
\title{This function find the post prob}
\usage{
find_postk(lambda, gamma, r, n, plo, phi)
}
\arguments{
\item{lambda}{the prior probability of H0: that the strata are completely correlated}

\item{gamma}{the prior probability that the drug is active in any specific stratum}

\item{r}{number of responses}

\item{n}{number of patients}

\item{plo}{response probability for inactive drug}

\item{phi}{response probability for active drug}
}
\value{
post prob of activity in each strata
}
\description{
This function find the post prob
}
