% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Optimal_2stage_design.R
\name{fixed_twostage}
\alias{fixed_twostage}
\title{Optim paramerters setting}
\usage{
fixed_twostage(
  K,
  Alpha,
  p0,
  p1,
  a0range = seq(alpha, 0.3, by = 0.1),
  b0range = seq(0.05, 0.3, by = 0.1),
  n12pts,
  nsim = 10000,
  track = TRUE
)
}
\arguments{
\item{K}{number of indications}

\item{Alpha}{global type I error}

\item{p0}{probability vector of the null hypothesis}

\item{p1}{probability vector of the alternative hypothesis}

\item{a0range}{range for grid search}

\item{b0range}{range for grid search}

\item{n12pts}{number of patients}

\item{nsim}{number of simulation used for type I and type II calculation}

\item{track}{tracking}
}
\value{
optim paramerters setting for optimal one stage design
}
\description{
Optim paramerters setting
}
