% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:FWER.R
\name{fwer}
\alias{fwer}
\title{Get FWER}
\usage{
fwer(post.p, type_one)
}
\arguments{
\item{post.p}{posterior probability of efficacy}

\item{type_one}{fixed type one error for each cohort}
}
\value{
family wise error rate for fixed type one error
}
\description{
Get FWER
}
