% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:EXNEX.R
\name{jags.exnex.post.samples}
\alias{jags.exnex.post.samples}
\title{Get posterior samples from MUCE Method}
\usage{
jags.exnex.post.samples(
  nk = rep(24, 6),
  rk = rep(4, 6),
  k = 6,
  Nexch = 2,
  Nmix = 3,
  pMix = c(0.5, 0, 0.5),
  dirichprior = c(1, 1, 1),
  dirichprior.check = TRUE,
  mu.mean = c(-1.735, 0.847),
  mu.prec = c(0.146, 0.266),
  tau.HN.scale = c(1, 1),
  nex.mean = rep(-1.734, 6),
  nex.prec = rep(0.128, 6),
  n.burnin = 1000,
  n.chain = 1,
  n.iter = 8000
)
}
\arguments{
\item{nk}{sample size for each cohort at each analysis (can be different). row is for analysis#}

\item{rk}{?}

\item{k}{no. of cohorts/indications}

\item{Nexch}{the no. of exchangeability distributions}

\item{Nmix}{==Nexch+1. no. of total candidate prior dists for theta_j.}

\item{pMix}{the probs of each strata following each dist.}

\item{dirichprior}{parameter for prior distribution specified in exnex.dirich.txt. but we do not have this file.}

\item{dirichprior.check}{if TRUE, will use exnex.dirich.txt to specify bayesian model}

\item{mu.mean}{mean of prior dist of mu in exchangeable part}

\item{mu.prec}{std of prior dist of mu in exchangeable part}

\item{tau.HN.scale}{parameter to derive prior dist of}

\item{nex.mean}{to derive prior dist of theta_j under non-exchangeability}

\item{nex.prec}{to derive prior dist of theta_j under non-exchangeability}

\item{n.burnin}{control prameters of MCMC sampling using jags}

\item{n.chain}{control prameters of MCMC sampling using jags}

\item{n.iter}{control prameters of MCMC sampling using jags}
}
\value{
Probability to stop
}
\description{
Get posterior samples from MUCE Method
}
