% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:MUCE.R
\name{jags.muce.post.samples}
\alias{jags.muce.post.samples}
\title{Get posterior samples from MUCE Method}
\usage{
jags.muce.post.samples(
  k,
  J,
  Nk,
  rik,
  H0,
  gamma,
  prec0,
  prec.eta,
  prec.ksi,
  mu.eta0,
  prec.eta0,
  mu.ksi0,
  prec.ksi0,
  n.burnin,
  n.chain,
  n.iter
)
}
\arguments{
\item{k}{no. of indications}

\item{J}{no. of dose levels}

\item{Nk}{? sample size at each analysis}

\item{rik}{?}

\item{H0}{largest response rate under null, of length(K)}

\item{gamma}{a parameter for prior dist of theta_ij}

\item{prec0}{std of Z_ij, latent probit regression}

\item{prec.eta}{std of dose specific effects}

\item{prec.ksi}{std of indication specific effects}

\item{mu.eta0}{the mean of hyperprior eta_0}

\item{prec.eta0}{the std of hyperprior eta_0}

\item{mu.ksi0}{the mean of hyperprior ksi_0}

\item{prec.ksi0}{the std of hyperprior ksi_0}

\item{n.burnin}{control prameters of MCMC sampling using jags}

\item{n.chain}{control prameters of MCMC sampling using jags}

\item{n.iter}{control prameters of MCMC sampling using jags}
}
\value{
Probability to stop
}
\description{
Get posterior samples from MUCE Method
}
