% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:MUCE.R
\name{muce.calibrate}
\alias{muce.calibrate}
\title{Get efficacy cutoff}
\usage{
muce.calibrate(
  K,
  J,
  L,
  num.sim,
  FWER,
  precision,
  fwer.threshold,
  nik,
  H0,
  H1,
  p0,
  v1,
  gamma,
  prec0,
  prec.eta,
  prec.ksi,
  mu.eta0,
  prec.eta0,
  mu.ksi0,
  prec.ksi0,
  n.burnin,
  n.chain,
  n.iter
)
}
\arguments{
\item{J}{no. of dose levels}

\item{L}{no. of interim analysis}

\item{num.sim}{no. of repetitions}

\item{FWER}{the pre-specified family wise error rate for efficacy cutoff calculatio}

\item{fwer.threshold}{the acceptable thresholf for the difference between estimated FWER and proposed FWER}

\item{nik}{sample size at each analysis}

\item{H0}{largest response rate under null, of length(K)}

\item{H1}{lowest response rate under alternative, of length(K)}

\item{p0}{true response rate for each arm}

\item{v1}{phi_1, the futility boundary at interim}

\item{gamma}{a parameter for prior dist of theta_ij}

\item{prec0}{std of Z_ij, latent probit regression}

\item{prec.eta}{std of dose specific effects}

\item{prec.ksi}{std of indication specific effects}

\item{mu.eta0}{the mean of hyperprior eta_0}

\item{prec.eta0}{the std of hyperprior eta_0}

\item{mu.ksi0}{the mean of hyperprior ksi_0}

\item{prec.ksi0}{the std of hyperprior ksi_0}

\item{n.burnin}{control prameters of MCMC sampling using jags}

\item{n.chain}{control prameters of MCMC sampling using jags}

\item{n.iter}{control prameters of MCMC sampling using jags}

\item{k}{no. of indications}

\item{precisio}{the grid space controlled in the grid search, can be refined if higher precision is needed}
}
\value{
efficacy cutoff
}
\description{
Get efficacy cutoff
}
