% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:MUCE.R
\name{muce}
\alias{muce}
\title{MUCE() design Function}
\usage{
muce(
  K,
  J,
  L,
  nik,
  H0,
  H1,
  p0,
  v1,
  v2,
  gamma,
  prec0,
  prec.eta,
  prec.ksi,
  mu.eta0,
  prec.eta0,
  mu.ksi0,
  prec.ksi0,
  n.burnin,
  n.chain,
  n.iter,
  r.all,
  num.sim = 1
)
}
\arguments{
\item{J}{no. of dose levels}

\item{L}{no. of interim analysis;}

\item{nik}{sample size at each analysis}

\item{H0}{largest response rate under null, of length(K)}

\item{H1}{lowest response rate under alternative, of length(K)}

\item{p0}{true response rate for each arm}

\item{v1}{phi_1, the futility boundary at interim}

\item{v2}{phi_2, the efficacy boundary at final}

\item{gamma}{a parameter for prior dist of theta_ij}

\item{prec0}{std of Z_ij, latent probit regression}

\item{prec.eta}{std of dose specific effects}

\item{prec.ksi}{std of indication specific effects}

\item{mu.eta0}{the mean of hyperprior eta_0}

\item{prec.eta0}{the std of hyperprior eta_0}

\item{mu.ksi0}{the mean of hyperprior ksi_0}

\item{prec.ksi0}{the std of hyperprior ksi_0}

\item{n.burnin}{control prameters of MCMC sampling using jags}

\item{n.chain}{control prameters of MCMC sampling using jags}

\item{n.iter}{control prameters of MCMC sampling using jags}

\item{r.all}{number of responses}

\item{num.sim}{number of simulated trials needs to be 1 here, repeated simulations are done outside}

\item{k}{no. of indications}
}
\value{
probability of reject
}
\description{
MUCE() design Function
}
