% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Optimal_1stage_design.R
\name{optimal.1stage}
\alias{optimal.1stage}
\title{main function decision making}
\usage{
optimal.1stage(n, k, r.1stg, alpha.adj.1stg, r, resp.cntl)
}
\arguments{
\item{n}{number of patients}

\item{k}{number of indications}

\item{r.1stg}{number of responses for interium 1st stage}

\item{alpha.adj.1stg}{adjusted alpha for 1st stage}

\item{resp.cntl}{probability under the null}
}
\value{
return drug active ones, total number of subjects
}
\description{
main function decision making
}
