% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Optimal_2stage_design.R
\name{Optimal.2stg}
\alias{Optimal.2stg}
\title{main function decision making}
\usage{
Optimal.2stg(k, n, r.stg1, n.stg1, alphastar.2stg, r.ia.2stg, r, resp.cntl)
}
\arguments{
\item{k}{number of indications}

\item{n}{number of patients}

\item{r.stg1}{number of responses for interium 1st stage}

\item{n.stg1}{number of responses for interium 1st stage}

\item{alphastar.2stg}{adjusted alpha for 2stg}

\item{r.ia.2stg}{number of responses for interium 2stg}

\item{r}{number of responses}

\item{resp.cntl}{probability under the null}
}
\value{
return drug active ones, total number of subjects
}
\description{
main function decision making
}
