% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Optimal_1stage_design.R
\name{Optimal_pooling_hetero}
\alias{Optimal_pooling_hetero}
\title{Optim paramerters setting}
\usage{
Optimal_pooling_hetero(K, n, p0, p1, Alpha, nsim = 1e+05)
}
\arguments{
\item{K}{number of indications}

\item{n}{number of patients}

\item{p0}{probability vector of the null hypothesis}

\item{p1}{probability vector of the alternative hypothesis}

\item{Alpha}{global type I error}

\item{nsim}{number of simulation used for type I and type II calculation}
}
\value{
optim paramerters setting for optimal one stage design
}
\description{
Optim paramerters setting
}
