% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Simon_two_stage.R
\name{PET}
\alias{PET}
\title{Probability to stop for futility at IA}
\usage{
PET(r1, n1, p)
}
\arguments{
\item{r1}{number of responses at IA}

\item{size}{sample size}

\item{prob}{probability vector}
}
\value{
Probability to stop
}
\description{
Probability to stop for futility at IA
}
