% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Optimal_2stage_design.R
\name{power.cal.2stg}
\alias{power.cal.2stg}
\title{This function finds alphastar}
\usage{
power.cal.2stg(alphastar, alpbeta0, n12pts, K, p0, p1, nsim, alpha)
}
\arguments{
\item{alphastar}{adjusted alpha}

\item{alpbeta0}{alpha}

\item{n12pts}{total sample size for each indication}

\item{K}{number of indications}

\item{p0}{probability vector of the null hypothesis}

\item{p1}{probability vector of the alternative hypothesis}

\item{nsim}{number of simulation used for type I and type II calculation}

\item{alpha}{type I error}
}
\value{
return alphastar
}
\description{
This function finds alphastar
}
