% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Optimal_1stage_design.R
\name{PowerEmpirical.1stg}
\alias{PowerEmpirical.1stg}
\title{This function finds alphastar}
\usage{
PowerEmpirical.1stg(K, g, r, n, p0, p1, alphastar, nsim)
}
\arguments{
\item{K}{number of indications}

\item{g}{the prior probability of H0: that the strata are completely correlated}

\item{r}{number of responses}

\item{n}{number of patients}

\item{p0}{probability vector of the null hypothesis}

\item{p1}{probability vector of the alternative hypothesis}

\item{alphastar}{adjusted alpha}

\item{nsim}{number of simulation used for type I and type II calculation}
}
\value{
power
}
\description{
This function finds alphastar
}
