% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Optimal_2stage_design.R
\name{PowerEmpirical.2stg}
\alias{PowerEmpirical.2stg}
\title{This function uses simulation to calculate the power}
\usage{
PowerEmpirical.2stg(
  K,
  g,
  alpha0,
  beta0,
  n12pts,
  alphastar,
  p0,
  p1,
  nsim = 10000
)
}
\arguments{
\item{K}{number of indications}

\item{g}{the prior probability of H0: that the strata are completely correlated}

\item{alpha0}{the type I error parameter for stage1}

\item{beta0}{the type II error parameter for stage1}

\item{n12pts}{total sample size for each indication}

\item{alphastar}{adjusted alpha}

\item{p0}{probability vector of the null hypothesis}

\item{p1}{probability vector of the alternative hypothesis}

\item{nsim}{number of simulation used for type I and type II calculation}
}
\value{
power
}
\description{
This function uses simulation to calculate the power
}
