% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Optimal_2stage_design.R
\name{PowerEmpirical.helper.2stg}
\alias{PowerEmpirical.helper.2stg}
\title{This helper function to calculate the power}
\usage{
PowerEmpirical.helper.2stg(K, g, r1, n1, n2, p0, p1, alphastar, nsim = 10000)
}
\arguments{
\item{K}{number of indications}

\item{g}{the prior probability of H0: that the strata are completely correlated}

\item{r1}{number of responses for stage I}

\item{n1}{number of patients for stage I}

\item{n2}{number of patients for stage II}

\item{p0}{probability vector of the null hypothesis}

\item{p1}{probability vector of the alternative hypothesis}

\item{alphastar}{adjusted alpha}

\item{nsim}{number of simulation used for type I and type II calculation}
}
\value{
power
}
\description{
This helper function to calculate the power
}
