% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Simon_two_stage.R
\name{ProbRej}
\alias{ProbRej}
\title{Prob for stop futility at end of the trial}
\usage{
ProbRej(r1, n1, p, r, n)
}
\arguments{
\item{r1}{number of responses at IA}

\item{n1}{number of patients at IA}

\item{p}{probability vector}

\item{r}{number of responses}

\item{n}{number of patients}
}
\value{
probability of reject
}
\description{
Prob for stop futility at end of the trial
}
