% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Simon_two_stage.R
\name{Simon.2stage}
\alias{Simon.2stage}
\title{main function decision making}
\usage{
Simon.2stage(n1.s2stg, n.s2stg, r1.s2stg, r.s2stg, r.ia, r)
}
\arguments{
\item{n1.s2stg}{number of patients}

\item{n.s2stg}{number of patients at 2stgs}

\item{r1.s2stg}{number of responses}

\item{r.s2stg}{number of responses}

\item{r.ia}{number of responses at IA}

\item{r}{number of responses}
}
\value{
return drug active ones, total number of subjects
}
\description{
main function decision making
}
