% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Simon_two_stage.R
\name{Simon.2stg.search}
\alias{Simon.2stg.search}
\title{Optim parameter}
\usage{
Simon.2stg.search(n.s2stg, n1.s2stg, p0.s2stg, p1.s2stg, r1.s2stg, r.s2stg)
}
\arguments{
\item{n.s2stg}{number of patients at 2stgs}

\item{n1.s2stg}{number of patients}

\item{p0.s2stg}{number of responses for interium 1st stage}

\item{p1.s2stg}{adjusted alpha for 1st stage}

\item{r1.s2stg}{number of responses}

\item{r.s2stg}{number of responses}
}
\value{
parameters
}
\description{
Optim parameter
}
