% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Simon_Bayes_design.R
\name{SimonBayes.2stg}
\alias{SimonBayes.2stg}
\title{main function decision making}
\usage{
SimonBayes.2stg(
  r.ia = NULL,
  r,
  lambda,
  gamma,
  plo,
  phi,
  T = 0.8,
  n.ia = NULL,
  k,
  n
)
}
\arguments{
\item{r.ia}{number of responses in interium analysis}

\item{r}{number of responses}

\item{lambda}{the prior probability of H0: that the strata are completely correlated}

\item{gamma}{the prior probability that the drug is active in any specific stratum}

\item{plo}{response probability for inactive drug}

\item{phi}{response probability for active drug}

\item{T}{threshold for conclusive posterior probability}

\item{n.ia}{number of patients for interium analysis}

\item{k}{number of cohorts}

\item{n}{Path to the input file}
}
\value{
return drug active ones, total number of subjects
}
\description{
main function decision making
}
