% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Simon_Bayes_design.R
\name{SimonBayes.type1}
\alias{SimonBayes.type1}
\title{Simulate data and run main function decision making}
\usage{
SimonBayes.type1(
  rep,
  k,
  n,
  n.ia,
  resp.true,
  resp.cntl,
  resp.trt,
  lambda,
  gamma,
  plo,
  phi,
  T
)
}
\arguments{
\item{rep}{number of repetition}

\item{k}{number of cohorts}

\item{n}{number of patients}

\item{n.ia}{number of patients for interium analysis}

\item{resp.true}{probability under the alternative}

\item{resp.cntl}{probability under the null}

\item{resp.trt}{probability under the alternative}

\item{lambda}{the prior probability of H0: that the strata are completely correlated}

\item{gamma}{the prior probability that the drug is active in any specific stratum}

\item{plo}{response probability for inactive drug}

\item{phi}{response probability for active drug}

\item{T}{threshold for conclusive posterior probability}
}
\value{
return drug active ones, total number of subjects
}
\description{
Simulate data and run main function decision making
}
