% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Simple_pooling.R
\name{Simple.Pool}
\alias{Simple.Pool}
\title{main function decision making}
\usage{
Simple.Pool(n, k, r, resp.cntl, alpha)
}
\arguments{
\item{n}{number of patients}

\item{r}{number of responses}

\item{resp.cntl}{probability under the null}

\item{alpha}{global type I error}

\item{K}{number of indications}
}
\value{
return drug active ones, total number of subjects
}
\description{
main function decision making
}
