% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Summary_results.R
\name{summary_sim}
\alias{summary_sim}
\title{Summary Results}
\usage{
summary_sim(
  res,
  p = 5,
  q = 4,
  mnames = c("SimonBayes", "Opt2stg", "Opt1stg", "Simon2stg", "Pool"),
  tp
)
}
\arguments{
\item{res}{results from Monte Carlo Simulation}

\item{p}{is the number of methods}

\item{q}{is the number of cohorts}

\item{mnames}{names of methods}

\item{tp}{True positive indicator}
}
\value{
summary results
}
\description{
Summary Results
}
