% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R:Simon_Bayes_design.R
\name{SimonBayes.type1.search}
\alias{SimonBayes.type1.search}
\title{search for typeI error under differen T}
\usage{
SimonBayes.type1.search(
  k,
  n,
  n.ia,
  resp.true,
  lambda,
  gamma,
  plo,
  phi,
  T,
  nsim
)
}
\arguments{
\item{k}{number of cohorts}

\item{n}{number of patients}

\item{n.ia}{number of patients for interium analysis}

\item{resp.true}{probability under the null}

\item{lambda}{the prior probability of H0: that the strata are completely correlated}

\item{gamma}{the prior probability that the drug is active in any specific stratum}

\item{plo}{response probability for inactive drug}

\item{phi}{response probability for active drug}

\item{T}{threshold for conclusive posterior probability}

\item{nsim}{Number of simulations for each T}
}
\value{
return typeI error under differen T
}
\description{
search for typeI error under differen T
}
